backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
openai
pydantic
python-dotenv
httpx
